requirements.txt
python-telegram-bot==13.5
requests==2.26.0
beautifulsoup4==4.9.3
web3==5.26.0
